+++ src/pipes/index.ts
@@ -1,0 +1,4 @@
+export * from './urlEncode/urlEncode.pipe';
+export * from './isPresent/isPresent.pipe';
+export * from './isNaN/isNaN.pipe';
+export * from './localize/localize.pipe';